Made an SVG folder and moved SVG components into it
